modify uint8_t to int32_t

+++ camera/vdi_base/usb_camera/adapter/platform/v4l2/src/device_manager/v4l2_device_manager.cpp
@@ -752,7 +752,7 @@ void V4L2DeviceManager::AddDefaultAbilityStreamAvailableBasicConfigurations(std:
 
 void V4L2DeviceManager::AddDefaultSensorOrientation(std::shared_ptr<CameraMetadata> metadata)
 {
-    std::vector<uint8_t> sensorOrientationVector;
+    std::vector<int32_t> sensorOrientationVector;
     const int32_t sensorOrientation = 0;
     sensorOrientationVector.push_back(sensorOrientation);
     AddOrUpdateOhosTag(metadata, OHOS_SENSOR_ORIENTATION, sensorOrientationVector);
